sdk/python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "polos-worker"
version = "0.1.0"
description = "Polos worker for Python - Durable Agent Execution"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.24.0",
    "pydantic>=2.0.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.20.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "anyio>=4.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
anthropic = [
    "anthropic>=0.39.0",
]
openai = [
    "openai>=1.0.0",
]
gemini = [
    "openai>=1.0.0",
]
groq = [
    "openai>=1.0.0",
]
fireworks = [
    "openai>=1.0.0",
]
together = [
    "openai>=1.0.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "ruff>=0.1.0",
    "pre-commit>=3.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["polos"]

[tool.ruff]
# Line length matches project style
line-length = 100
target-version = "py310"

# Enable formatting
[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

# Linting rules
[tool.ruff.lint]
# Enable specific rule sets
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort (import sorting)
    "N",   # pep8-naming
    "UP",  # pyupgrade (modernize code)
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]

# Ignore specific rules if needed
ignore = []

# Allow autofix for all enabled rules
fixable = ["ALL"]
unfixable = []

# Import sorting configuration
[tool.ruff.lint.isort]
known-first-party = ["polos"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "ruff>=0.1.0",
    "pre-commit>=3.0.0",
]